refactor(game): update difficulty tiers and reward types
Replaced the EXTREME difficulty with MASTER, GRANDMASTER, and LEGENDARY tiers. Updated reward currency type to Double for increased precision.

+++ android/app/src/main/java/com/arrowescape/game/engine/LevelGenerator.kt
@@ -383,7 +383,7 @@ object LevelGenerator {
         levelId: Int,
         name: String,
         difficulty: Difficulty,
-        rewardRupees: Int,
+        rewardRupees: Double,
         gridWidth: Int,
         gridHeight: Int,
         targetCount: Int
